+++ frontend/src/page/exercise.rs
@@ -22,25 +22,24 @@ pub fn init(
         .unwrap_or("")
         .parse::<u32>()
         .unwrap_or(0);
+    let editing = url.next_hash_path_part() == Some("edit");
 
     orders.subscribe(Msg::DataEvent);
 
     navbar.title = String::from("Exercise");
 
-    Model {
-        interval: common::init_interval(
-            &data_model
-                .training_sessions
-                .values()
-                .filter(|t| t.exercises().contains(&exercise_id))
-                .map(|t| t.date)
-                .collect::<Vec<NaiveDate>>(),
-            false,
-        ),
+    let mut model = Model {
+        interval: common::init_interval(&[], true),
         exercise_id,
+        name: InputField::default(),
         dialog: Dialog::Hidden,
+        editing,
         loading: false,
-    }
+    };
+
+    update_model(&mut model, data_model);
+
+    model
 }
 
 // ------ ------
@@ -50,8 +49,25 @@ pub fn init(
 pub struct Model {
     interval: common::Interval,
     exercise_id: u32,
+    name: InputField<String>,
     dialog: Dialog,
+    editing: bool,
     loading: bool,
+}
+
+impl Model {
+    pub fn has_unsaved_changes(&self) -> bool {
+        self.name.changed()
+    }
+
+    pub fn mark_as_unchanged(&mut self) {
+        self.name.input = self.name.parsed.clone().unwrap();
+        self.name.orig = self.name.parsed.clone().unwrap();
+    }
+
+    fn saving_disabled(&self) -> bool {
+        self.loading || not(self.name.valid())
+    }
 }
 
 enum Dialog {
@@ -59,13 +75,45 @@ enum Dialog {
     DeleteTrainingSession(u32),
 }
 
+#[derive(Clone)]
+struct InputField<T> {
+    input: String,
+    parsed: Option<T>,
+    orig: String,
+}
+
+impl<T> Default for InputField<T> {
+    fn default() -> Self {
+        InputField {
+            input: String::new(),
+            parsed: None,
+            orig: String::new(),
+        }
+    }
+}
+
+impl<T> InputField<T> {
+    fn valid(&self) -> bool {
+        self.parsed.is_some()
+    }
+
+    fn changed(&self) -> bool {
+        self.input != self.orig
+    }
+}
+
 // ------ ------
 //    Update
 // ------ ------
 
 pub enum Msg {
+    EditExercise,
+    SaveExercise,
+
     ShowDeleteTrainingSessionDialog(u32),
     CloseDialog,
+
+    NameChanged(String),
 
     DeleteTrainingSession(u32),
     DataEvent(data::Event),
@@ -80,6 +128,23 @@ pub fn update(
     orders: &mut impl Orders<Msg>,
 ) {
     match msg {
+        Msg::EditExercise => {
+            model.editing = true;
+            Url::go_and_push(
+                &crate::Urls::new(&data_model.base_url)
+                    .exercise()
+                    .add_hash_path_part(model.exercise_id.to_string())
+                    .add_hash_path_part("edit"),
+            );
+        }
+        Msg::SaveExercise => {
+            model.loading = true;
+            orders.notify(data::Msg::ReplaceExercise(data::Exercise {
+                id: model.exercise_id,
+                name: model.name.parsed.clone().unwrap(),
+            }));
+        }
+
         Msg::ShowDeleteTrainingSessionDialog(position) => {
             model.dialog = Dialog::DeleteTrainingSession(position);
         }
@@ -93,6 +158,28 @@ pub fn update(
             );
         }
 
+        Msg::NameChanged(name) => {
+            let trimmed_name = name.trim();
+            if not(trimmed_name.is_empty())
+                && (trimmed_name == model.name.orig
+                    || data_model
+                        .exercises
+                        .values()
+                        .all(|e| e.name != trimmed_name))
+            {
+                model.name = InputField {
+                    input: name.clone(),
+                    parsed: Some(trimmed_name.to_string()),
+                    orig: model.name.orig.clone(),
+                };
+            } else {
+                model.name = InputField {
+                    input: name,
+                    parsed: None,
+                    orig: model.name.orig.clone(),
+                }
+            }
+        }
         Msg::DeleteTrainingSession(id) => {
             model.loading = true;
             orders.notify(data::Msg::DeleteTrainingSession(id));
@@ -101,14 +188,15 @@ pub fn update(
             model.loading = false;
             match event {
                 data::Event::DataChanged => {
-                    model.interval = common::init_interval(
-                        &data_model
-                            .training_sessions
-                            .values()
-                            .filter(|t| t.exercises().contains(&model.exercise_id))
-                            .map(|t| t.date)
-                            .collect::<Vec<NaiveDate>>(),
-                        false,
+                    update_model(model, data_model);
+                }
+                data::Event::ExerciseReplacedOk => {
+                    model.editing = false;
+                    model.mark_as_unchanged();
+                    Url::go_and_push(
+                        &crate::Urls::new(&data_model.base_url)
+                            .exercise()
+                            .add_hash_path_part(model.exercise_id.to_string()),
                     );
                 }
                 data::Event::TrainingSessionDeletedOk => {
@@ -125,6 +213,28 @@ pub fn update(
     }
 }
 
+fn update_model(model: &mut Model, data_model: &data::Model) {
+    model.interval = common::init_interval(
+        &data_model
+            .training_sessions
+            .values()
+            .filter(|t| t.exercises().contains(&model.exercise_id))
+            .map(|t| t.date)
+            .collect::<Vec<NaiveDate>>(),
+        false,
+    );
+
+    let exercise = &data_model.exercises.get(&model.exercise_id);
+
+    if let Some(exercise) = exercise {
+        model.name = InputField {
+            input: exercise.name.clone(),
+            parsed: Some(exercise.name.clone()),
+            orig: exercise.name.clone(),
+        };
+    };
+}
+
 // ------ ------
 //     View
 // ------ ------
@@ -132,7 +242,7 @@ pub fn update(
 pub fn view(model: &Model, data_model: &data::Model) -> Node<Msg> {
     if data_model.exercises.is_empty() && data_model.loading_exercises {
         common::view_page_loading()
-    } else if let Some(exercise) = data_model.exercises.get(&model.exercise_id) {
+    } else if model.exercise_id > 0 {
         let exercise_training_sessions = exercise_training_sessions(model, data_model);
         let dates = exercise_training_sessions.iter().map(|t| t.date);
         let exercise_interval = common::Interval {
@@ -144,21 +254,78 @@ pub fn view(model: &Model, data_model: &data::Model) -> Node<Msg> {
             .filter(|t| t.date >= model.interval.first && t.date <= model.interval.last)
             .collect::<Vec<_>>();
         div![
-            common::view_title(&span![&exercise.name], 5),
-            common::view_interval_buttons(&model.interval, &exercise_interval, Msg::ChangeInterval),
-            view_charts(&training_sessions, &model.interval),
-            view_calendar(&training_sessions, &model.interval),
-            training::view_table(
-                &training_sessions,
-                &data_model.routines,
-                &data_model.base_url,
-                Msg::ShowDeleteTrainingSessionDialog
-            ),
-            view_dialog(&model.dialog, model.loading)
+            view_title(model),
+            if model.editing {
+                nodes![button![
+                    C!["button"],
+                    C!["is-fab"],
+                    C!["is-medium"],
+                    C!["is-link"],
+                    C![IF![model.loading => "is-loading"]],
+                    attrs![
+                        At::Disabled => model.saving_disabled().as_at_value(),
+                    ],
+                    ev(Ev::Click, |_| Msg::SaveExercise),
+                    span![C!["icon"], i![C!["fas fa-save"]]]
+                ]]
+            } else {
+                nodes![
+                    common::view_interval_buttons(
+                        &model.interval,
+                        &exercise_interval,
+                        Msg::ChangeInterval
+                    ),
+                    view_charts(&training_sessions, &model.interval),
+                    view_calendar(&training_sessions, &model.interval),
+                    training::view_table(
+                        &training_sessions,
+                        &data_model.routines,
+                        &data_model.base_url,
+                        Msg::ShowDeleteTrainingSessionDialog
+                    ),
+                    view_dialog(&model.dialog, model.loading),
+                    common::view_fab("edit", |_| Msg::EditExercise),
+                ]
+            },
         ]
     } else {
         common::view_error_not_found("Exercise")
     }
+}
+
+fn view_title(model: &Model) -> Node<Msg> {
+    div![
+        C!["mx-2"],
+        C!["mb-5"],
+        if model.editing {
+            let saving_disabled = model.saving_disabled();
+            div![
+                C!["field"],
+                div![
+                    C!["control"],
+                    input_ev(Ev::Input, Msg::NameChanged),
+                    keyboard_ev(Ev::KeyDown, move |keyboard_event| {
+                        IF!(
+                            not(saving_disabled) && keyboard_event.key_code() == common::ENTER_KEY => {
+                                Msg::SaveExercise
+                            }
+                        )
+                    }),
+                    input![
+                        C!["input"],
+                        C![IF![not(model.name.valid()) => "is-danger"]],
+                        C![IF![model.name.changed() => "is-info"]],
+                        attrs! {
+                            At::Type => "text",
+                            At::Value => model.name.input,
+                        }
+                    ]
+                ],
+            ]
+        } else {
+            common::view_title(&span![&model.name.input], 0)
+        }
+    ]
 }
 
 pub fn view_charts<Ms>(