+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Prediction of next menstrual cycle
 - Calculation of average and variation of menstrual cycle length
 - Search box to exercises and routines page
-- Renaming on routine page
+- Renaming on exercise and routine page
 - Warning about unsaved changes before leaving page
 - Possibility to create exercises while editing routines
 - Loading indicator to all pages